move moduleShortMapper to package module
