+++ patient/src/main/java/ru/rightcode/patient/mapper/module/ModuleShortMapper.java
@@ -1,10 +1,9 @@
-package ru.rightcode.patient.mapper.rehab;
+package ru.rightcode.patient.mapper.module;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import ru.rightcode.patient.dto.response.rehab.ModuleShortResponse;
-import ru.rightcode.patient.mapper.module.ExerciseShortResponseMapper;
 import ru.rightcode.patient.model.Module;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING,
