+++ patient/src/main/java/ru/rightcode/patient/mapper/rehab/RehabProgramMapper.java
@@ -4,6 +4,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingConstants;
 import ru.rightcode.patient.dto.response.rehab.RehabProgramResponse;
+import ru.rightcode.patient.mapper.module.ModuleShortMapper;
 import ru.rightcode.patient.model.RehabProgram;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING,
